requirements.txt
adal==1.2.7
argcomplete==2.0.0
azure-common==1.1.28
azure-core==1.24.2
azure-graphrbac==0.61.1
azure-mgmt-authorization==2.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==10.0.0
azure-mgmt-resource==21.1.0
azure-mgmt-storage==20.0.0
azure-storage-blob==12.13.0
azure-storage-file-datalake==12.8.0
azureml-core==1.43.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.2
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
contextlib2==21.6.0
cryptography==37.0.4
docker==5.0.3
humanfriendly==10.0
idna==3.3
isodate==0.6.1
jeepney==0.8.0
jmespath==1.0.0
jsonpickle==2.2.0
knack==0.9.0
msal==1.18.0
msal-extensions==1.0.0
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
oauthlib==3.2.0
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
pkginfo==1.8.3
portalocker==2.5.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
SecretStorage==3.3.2
six==1.16.0
tabulate==0.8.10
typing_extensions==4.3.0
urllib3==1.26.9
websocket-client==1.3.3
